Move OciWrapper class in its own file.
